# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Payment Gateway API",
        "description": "Payment Gateway API for processing payments, managing refunds, and handling webhooks",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Current Host"
        }
    ],
    "paths": {
...
        "/api/apps/{appId}/payments/initialize": {
            "post": {
                "tags": [
                    "App Payments"
                ],
                "summary": "Initialize an app payment",
                "description": "Initialize a payment on behalf of an app + an app user.",
                "operationId": "9286f9e465d36c6dca8226f0e70d2b69",
                "parameters": [
                    {
                        "name": "appId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "amount",
                                    "currency",
                                    "app_user"
                                ],
                                "properties": {
                                    "amount": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 250
                                    },
                                    "currency": {
                                        "type": "string",
                                        "example": "ZMW"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Top-up"
                                    },
                                    "callback_url": {
                                        "type": "string",
                                        "example": "https://yourapp.com/callback"
                                    },
                                    "app_user": {
                                        "required": [
                                            "external_user_id"
                                        ],
                                        "properties": {
                                            "external_user_id": {
                                                "type": "string",
                                                "example": "user_42"
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "Jane Doe"
                                            },
                                            "email": {
                                                "type": "string",
                                                "example": "jane@example.com"
                                            },
                                            "phone": {
                                                "type": "string",
                                                "example": "260970000000"
                                            }
                                        },
                                        "type": "object"
                                    },
                                    "metadata": {
                                        "type": "object",
                                        "example": {
                                            "order_id": "123"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Payment initialized successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "reference": {
                                            "type": "string"
                                        },
                                        "pay_url": {
                                            "type": "string"
                                        },
                                        "amount": {
                                            "type": "number",
                                            "format": "float"
                                        },
                                        "currency": {
                                            "type": "string"
                                        },
                                        "status": {
                                            "type": "string"
                                        },
                                        "app_user": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "apiKey": []
                    }
                ]
            }
        },
        "/api/apps/{appId}/payments": {
            "get": {
                "tags": [
                    "App Payments"
                ],
                "summary": "List app payments",
                "description": "List all payments for a specific app.",
                "operationId": "5157bbc4b58fe95153e2ea3541df560a",
                "parameters": [
                    {
                        "name": "appId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "external_user_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of payments retrieved successfully"
                    },
                    "404": {
                        "description": "App not found"
                    }
                },
                "security": [
                    {
                        "apiKey": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "apiKey": {
                "type": "http",
                "bearerFormat": "API Key",
                "scheme": "bearer"
            }
        }
    },
...
}